Monday, September 4, 2023 at 1:57 PM West Africa Standard Time
