+++ app/lib/pngCategories.js
@@ -1,8 +1,4 @@
-function getRandomPNGUrl(categoryName, gender) {
-  console.log(categoryName);
-  console.log(gender);
-
-  const categoryToNumberOfPngs = {
+const categoryToNumberOfPngs = {
     "adventurousFemale": 2,
     "adventurousMale": 2,
     "ambitiousMale": 2,
@@ -39,17 +35,4 @@ function getRandomPNGUrl(categoryName, gender) {
     "wittyMale" : 1,
   };
 
-  const numberOfPngs = categoryToNumberOfPngs[categoryName + gender] || 0;
-  if (numberOfPngs === 0) {
-    console.error(`No category found for: ${categoryName}`);
-    return null ;
-  }
-
-  const randomIndex = Math.floor(Math.random() * numberOfPngs) + 1;
-  const pngFileName = `${categoryName}${gender}_${randomIndex}.png`;
-
-  const pngUrl = `/assets/avatars/${categoryName}/${pngFileName}`;
-  return pngUrl;
-}
-
-export { getRandomPNGUrl };
+  export {categoryToNumberOfPngs}
