--- app/lib/roomIdGenerator.js
@@ -1,13 +1,0 @@
-function generateRoomId() {
-  let characters =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-
-  let peerId = "";
-
-  for (let i = 0; i < 10; i++)
-    peerId += characters.charAt(Math.floor(Math.random() * characters.length));
-
-  return peerId;
-}
-
-export { generateRoomId };
